requirements.txt
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.38.0
iniconfig==2.0.0
joblib==1.2.0
keras==2.11.0
libclang==16.0.6
littleutils==0.2.2
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
numpy==1.21.6
openpyxl==3.1.2
opt-einsum==3.3.0
outdated==0.2.2
packaging==23.0
pandas==1.3.5
pandas-flavor==0.5.0
patsy==0.5.3
Pillow==9.4.0
pingouin==0.5.3
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
SciencePlots==2.0.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
SimpleITK==2.2.1
sniffio==1.3.0
starlette==0.27.0
statsmodels==0.13.5
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.7.1
torchaudio==0.7.2
torchsummary==1.5.1
torchvision==0.8.2
urllib3==1.26.15
wrapt==1.15.0
xarray==0.20.2
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.15.0
